\name{runShiny}
\alias{runShiny}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{function to launch Shiny application for complexity function
}
\description{
Launches a Shiny application for the complexity function.
}
\usage{
runShiny()
}
\value{
A print of the following:
\item{true permutations}{ a print of the permutations in line with the constraints }
\item{total number of permutations}{ the total number of permutations }
\item{number true}{ the number of true permutations }
\item{complexity (proportion)}{ the complexity, that is: the proportion of true permutations }
}
\keyword{htest}
